Moving vsphere client creation to more appropriate package.
