+++ pkg/asset/installconfig/vsphere/client.go
@@ -2,10 +2,15 @@ package vsphere
 
 import (
 	"context"
+	"net/url"
+	"time"
 
+	"github.com/vmware/govmomi"
 	"github.com/vmware/govmomi/find"
 	"github.com/vmware/govmomi/object"
+	"github.com/vmware/govmomi/vapi/rest"
 	"github.com/vmware/govmomi/vim25"
+	"github.com/vmware/govmomi/vim25/soap"
 )
 
 // Finder interface represents the client that is used to connect to VSphere to get specific
@@ -28,4 +33,31 @@ type Finder interface {
 // vmware govmomi finder object that can be used to search for resources in vsphere.
 func NewFinder(client *vim25.Client, all ...bool) Finder {
 	return find.NewFinder(client, all...)
+}
+
+// CreateVSphereClients creates the SOAP and REST client to access
+// different portions of the vSphere API
+// e.g. tags are only available in REST
+func CreateVSphereClients(ctx context.Context, vcenter, username, password string) (*vim25.Client, *rest.Client, error) {
+	ctx, cancel := context.WithTimeout(ctx, 60*time.Second)
+	defer cancel()
+
+	u, err := soap.ParseURL(vcenter)
+	if err != nil {
+		return nil, nil, err
+	}
+	u.User = url.UserPassword(username, password)
+	c, err := govmomi.NewClient(ctx, u, false)
+
+	if err != nil {
+		return nil, nil, err
+	}
+
+	restClient := rest.NewClient(c.Client)
+	err = restClient.Login(ctx, u.User)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return c.Client, restClient, nil
 }
